+++ demo/app/pages/+default/home/home.component.ts
@@ -1,0 +1,21 @@
+import {Component, ChangeDetectionStrategy} from '@angular/core';
+import {ComponentRedirectRoute, ComponentRoute} from '@anglr/common/router';
+
+/**
+ * Home component
+ */
+@Component(
+{
+    selector: 'home-view',
+    templateUrl: 'home.component.html',
+    standalone: true,
+    imports:
+    [
+    ],
+    changeDetection: ChangeDetectionStrategy.OnPush
+})
+@ComponentRedirectRoute('', 'home')
+@ComponentRoute({path: 'home'})
+export class HomeComponent
+{
+}